move request, apiLooper, getUnixTimeStamp to common/utils
