+++ src/common/utils/getUnixTimeStamp.js
@@ -1,0 +1,3 @@
+const getUnixTimeStamp = () => Math.floor(new Date().getTime() / 1000)
+
+export default getUnixTimeStamp
